# backend/api/openapi/swagger.yaml
definitions:
...
  sico-backend_internal_transport_http_dto_rbac_casbin_rule.CasbinRule:
    properties:
      id:
        type: integer
      ptype:
        type: string
      v0:
        type: string
      v1:
        type: string
      v2:
        type: string
      v3:
        type: string
      v4:
        type: string
      v5:
        type: string
    required:
    - ptype
    - v0
    - v1
    - v2
    type: object
  sico-backend_internal_transport_http_dto_rbac_casbin_rule.CreatePolicyRequest:
    properties:
      ptype:
        type: string
      v0:
        type: string
      v1:
        type: string
      v2:
        type: string
      v3:
        type: string
      v4:
        type: string
      v5:
        type: string
    required:
    - ptype
    - v0
    - v1
    - v2
    type: object
  sico-backend_internal_transport_http_dto_rbac_casbin_rule.CreatePolicyResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/sico-backend_internal_transport_http_dto_rbac_casbin_rule.CreatePolicyResponseData'
      msg:
        type: string
    type: object
  sico-backend_internal_transport_http_dto_rbac_casbin_rule.CreatePolicyResponseData:
    properties:
      id:
        type: integer
    type: object
  sico-backend_internal_transport_http_dto_rbac_casbin_rule.DeletePolicyResponse:
    properties:
      code:
        type: integer
      msg:
        type: string
    type: object
  sico-backend_internal_transport_http_dto_rbac_casbin_rule.GetPolicyResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/sico-backend_internal_transport_http_dto_rbac_casbin_rule.GetPolicyResponseData'
      msg:
        type: string
    type: object
  sico-backend_internal_transport_http_dto_rbac_casbin_rule.GetPolicyResponseData:
    properties:
      policy:
        $ref: '#/definitions/sico-backend_internal_transport_http_dto_rbac_casbin_rule.CasbinRule'
    type: object
  sico-backend_internal_transport_http_dto_rbac_casbin_rule.QueryPoliciesResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/sico-backend_internal_transport_http_dto_rbac_casbin_rule.QueryPoliciesResponseData'
      msg:
        type: string
    type: object
  sico-backend_internal_transport_http_dto_rbac_casbin_rule.QueryPoliciesResponseData:
    properties:
      hasNext:
        type: boolean
      policies:
        items:
          $ref: '#/definitions/sico-backend_internal_transport_http_dto_rbac_casbin_rule.CasbinRule'
        type: array
      total:
        type: integer
    type: object
  sico-backend_internal_transport_http_dto_rbac_casbin_rule.UpdatePolicyRequest:
    properties:
      id:
        type: integer
      v0:
        type: string
      v1:
        type: string
      v2:
        type: string
      v3:
        type: string
      v4:
        type: string
      v5:
        type: string
    required:
    - id
    type: object
  sico-backend_internal_transport_http_dto_rbac_casbin_rule.UpdatePolicyResponse:
    properties:
      code:
        type: integer
      msg:
        type: string
    type: object
  sico-backend_internal_transport_http_dto_rbac_common.User:
    properties:
      alias:
        type: string
      createdAt:
        type: integer
      description:
        type: string
      email:
        type: string
      iconUri:
        type: string
      id:
        type: integer
      phone:
        type: string
      rawIconUri:
        type: string
      roles:
        items:
          type: string
        type: array
      status:
        $ref: '#/definitions/sico-backend_internal_transport_http_dto_rbac_common.UserStatus'
      tenant:
        type: string
      updatedAt:
        type: integer
      username:
        type: string
    type: object
  sico-backend_internal_transport_http_dto_rbac_common.UserStatus:
    enum:
    - 0
    - 1
    - 2
    format: int32
    type: integer
    x-enum-varnames:
    - UserStatus_USER_STATUS_INACTIVE
    - UserStatus_USER_STATUS_ACTIVE
    - UserStatus_USER_STATUS_FROZEN
...
  sico-backend_internal_transport_http_dto_rbac_token.LoginRequest:
    properties:
      email:
        maxLength: 64
        minLength: 3
        type: string
      password:
        maxLength: 128
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  sico-backend_internal_transport_http_dto_rbac_token.LoginResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/sico-backend_internal_transport_http_dto_rbac_token.LoginResponseData'
      msg:
        type: string
    type: object
  sico-backend_internal_transport_http_dto_rbac_token.LoginResponseData:
    properties:
      tokenInfo:
        $ref: '#/definitions/sico-backend_internal_transport_http_dto_rbac_token.TokenInfo'
      user:
        $ref: '#/definitions/sico-backend_internal_transport_http_dto_rbac_common.User'
    type: object
  sico-backend_internal_transport_http_dto_rbac_token.LogoutResponse:
    properties:
      code:
        type: integer
      msg:
        type: string
    type: object
  sico-backend_internal_transport_http_dto_rbac_token.RefreshTokenRequest:
    type: object
  sico-backend_internal_transport_http_dto_rbac_token.RefreshTokenResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/sico-backend_internal_transport_http_dto_rbac_token.RefreshTokenResponseData'
      msg:
        type: string
    type: object
  sico-backend_internal_transport_http_dto_rbac_token.RefreshTokenResponseData:
    properties:
      tokenInfo:
        $ref: '#/definitions/sico-backend_internal_transport_http_dto_rbac_token.TokenInfo'
    type: object
  sico-backend_internal_transport_http_dto_rbac_token.TokenInfo:
    properties:
      accessToken:
        type: string
      expiresAt:
        type: integer
      tokenType:
        type: string
    type: object
  sico-backend_internal_transport_http_dto_rbac_user.CreateUserRequest:
    properties:
      alias:
        maxLength: 64
        minLength: 1
        type: string
      description:
        maxLength: 1024
        type: string
      email:
        maxLength: 255
        type: string
      iconUri:
        maxLength: 255
        type: string
      password:
        maxLength: 128
        minLength: 6
        type: string
      phone:
        type: string
      tenant:
        maxLength: 255
        type: string
      username:
        maxLength: 64
        minLength: 3
        type: string
    required:
    - email
    - password
    type: object
  sico-backend_internal_transport_http_dto_rbac_user.CreateUserResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/sico-backend_internal_transport_http_dto_rbac_user.CreateUserResponseData'
      msg:
        type: string
    type: object
  sico-backend_internal_transport_http_dto_rbac_user.CreateUserResponseData:
    properties:
      id:
        type: integer
    type: object
  sico-backend_internal_transport_http_dto_rbac_user.DeleteUserResponse:
    properties:
      code:
        type: integer
      msg:
        type: string
    type: object
  sico-backend_internal_transport_http_dto_rbac_user.GetUserResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/sico-backend_internal_transport_http_dto_rbac_user.GetUserResponseData'
      msg:
        type: string
    type: object
  sico-backend_internal_transport_http_dto_rbac_user.GetUserResponseData:
    properties:
      user:
        $ref: '#/definitions/sico-backend_internal_transport_http_dto_rbac_common.User'
    type: object
  sico-backend_internal_transport_http_dto_rbac_user.QueryUsersResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/sico-backend_internal_transport_http_dto_rbac_user.QueryUsersResponseData'
      msg:
        type: string
    type: object
  sico-backend_internal_transport_http_dto_rbac_user.QueryUsersResponseData:
    properties:
      hasNext:
        type: boolean
      total:
        type: integer
      users:
        items:
          $ref: '#/definitions/sico-backend_internal_transport_http_dto_rbac_common.User'
        type: array
    type: object
  sico-backend_internal_transport_http_dto_rbac_user.ResetPasswordRequest:
    properties:
      id:
        type: integer
      newPassword:
        maxLength: 128
        minLength: 6
        type: string
      oldPassword:
        maxLength: 128
        minLength: 6
        type: string
    required:
    - id
    - newPassword
    - oldPassword
    type: object
  sico-backend_internal_transport_http_dto_rbac_user.ResetPasswordResponse:
    properties:
      code:
        type: integer
      msg:
        type: string
    type: object
  sico-backend_internal_transport_http_dto_rbac_user.UpdateUserRequest:
    properties:
      alias:
        maxLength: 64
        minLength: 1
        type: string
      description:
        maxLength: 1024
        type: string
      email:
        maxLength: 255
        type: string
      iconUri:
        maxLength: 255
        type: string
      id:
        type: integer
      phone:
        type: string
      status:
        allOf:
        - $ref: '#/definitions/sico-backend_internal_transport_http_dto_rbac_common.UserStatus'
        enum:
        - 0
        - 1
        - 2
      tenant:
        maxLength: 255
        type: string
    required:
    - id
    type: object
  sico-backend_internal_transport_http_dto_rbac_user.UpdateUserResponse:
    properties:
      code:
        type: integer
      msg:
        type: string
    type: object
...
info:
  contact: {}
  description: Sico AI Agent Platform API
  title: Sico API
  version: "1.0"
paths:
...
  /api/sico/rbac/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/sico-backend_internal_transport_http_dto_rbac_token.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sico-backend_internal_transport_http_dto_rbac_token.LoginResponse'
      tags:
      - RBAC
  /api/sico/rbac/logout:
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sico-backend_internal_transport_http_dto_rbac_token.LogoutResponse'
      security:
      - BearerAuth: []
      tags:
      - RBAC
  /api/sico/rbac/policies:
    get:
      consumes:
      - application/json
      parameters:
      - default: 1
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 10
        in: query
        maximum: 100
        minimum: 1
        name: pageSize
        type: integer
      - in: query
        name: ptype
        type: string
      - in: query
        name: v0
        type: string
      - in: query
        name: v1
        type: string
      - in: query
        name: v2
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sico-backend_internal_transport_http_dto_rbac_casbin_rule.QueryPoliciesResponse'
      security:
      - BearerAuth: []
      summary: Query Policies
      tags:
      - RBAC
  /api/sico/rbac/policy:
    delete:
      consumes:
      - application/json
      parameters:
      - in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sico-backend_internal_transport_http_dto_rbac_casbin_rule.DeletePolicyResponse'
      security:
      - BearerAuth: []
      summary: Delete Policy
      tags:
      - RBAC
    get:
      consumes:
      - application/json
      parameters:
      - in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sico-backend_internal_transport_http_dto_rbac_casbin_rule.GetPolicyResponse'
      security:
      - BearerAuth: []
      summary: Get Policy
      tags:
      - RBAC
    post:
      consumes:
      - application/json
      parameters:
      - description: Create Policy
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/sico-backend_internal_transport_http_dto_rbac_casbin_rule.CreatePolicyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sico-backend_internal_transport_http_dto_rbac_casbin_rule.CreatePolicyResponse'
      security:
      - BearerAuth: []
      summary: Create Policy
      tags:
      - RBAC
    put:
      consumes:
      - application/json
      parameters:
      - description: Update Policy
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/sico-backend_internal_transport_http_dto_rbac_casbin_rule.UpdatePolicyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sico-backend_internal_transport_http_dto_rbac_casbin_rule.UpdatePolicyResponse'
      security:
      - BearerAuth: []
      summary: Update Policy
      tags:
      - RBAC
  /api/sico/rbac/refresh:
    post:
      consumes:
      - application/json
      parameters:
      - description: Refresh token request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/sico-backend_internal_transport_http_dto_rbac_token.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sico-backend_internal_transport_http_dto_rbac_token.RefreshTokenResponse'
      security:
      - BearerAuth: []
      tags:
      - RBAC
...
  /api/sico/rbac/user:
    delete:
      consumes:
      - application/json
      parameters:
      - in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sico-backend_internal_transport_http_dto_rbac_user.DeleteUserResponse'
      security:
      - BearerAuth: []
      tags:
      - RBAC
    get:
      consumes:
      - application/json
      parameters:
      - in: query
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sico-backend_internal_transport_http_dto_rbac_user.GetUserResponse'
      security:
      - BearerAuth: []
      tags:
      - RBAC
    post:
      consumes:
      - application/json
      parameters:
      - description: Create user request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/sico-backend_internal_transport_http_dto_rbac_user.CreateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sico-backend_internal_transport_http_dto_rbac_user.CreateUserResponse'
      tags:
      - RBAC
    put:
      consumes:
      - application/json
      parameters:
      - description: Update user request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/sico-backend_internal_transport_http_dto_rbac_user.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sico-backend_internal_transport_http_dto_rbac_user.UpdateUserResponse'
      security:
      - BearerAuth: []
      tags:
      - RBAC
  /api/sico/rbac/user/password:
    put:
      consumes:
      - application/json
      parameters:
      - description: Reset password request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/sico-backend_internal_transport_http_dto_rbac_user.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sico-backend_internal_transport_http_dto_rbac_user.ResetPasswordResponse'
      security:
      - BearerAuth: []
      tags:
      - RBAC
...
  /api/sico/rbac/users:
    get:
      consumes:
      - application/json
      parameters:
      - in: query
        maxLength: 64
        name: alias
        type: string
      - in: query
        maxLength: 255
        name: email
        type: string
      - default: 1
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 10
        in: query
        maximum: 100
        minimum: 1
        name: pageSize
        type: integer
      - in: query
        maxLength: 32
        name: phone
        type: string
      - in: query
        name: statusList
        type: string
      - in: query
        name: tenant
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sico-backend_internal_transport_http_dto_rbac_user.QueryUsersResponse'
      security:
      - BearerAuth: []
      tags:
      - RBAC
...
securityDefinitions:
  BearerAuth:
    description: 'Enter your bearer token in the format: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"